+++ src/StateR/AfterEffects/IActionAfterEffects.cs
@@ -1,9 +1,9 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace StateR
+namespace StateR.AfterEffects
 {
-    public interface IAfterEffects<TAction>
+    public interface IActionAfterEffects<TAction>
         where TAction : IAction
     {
         Task HandleAfterEffectAsync(IDispatchContext<TAction> context, CancellationToken cancellationToken);
