+++ src/StateR/AfterEffects/IAfterEffectsManager.cs
@@ -1,0 +1,4 @@
+﻿namespace StateR.AfterEffects
+{
+    public interface IAfterEffectsManager : IDispatchManager { }
+}
