+++ src/StateR/DispatchContextFactory.cs
@@ -1,0 +1,9 @@
+﻿namespace StateR
+{
+    public class DispatchContextFactory : IDispatchContextFactory
+    {
+        public IDispatchContext<TAction> Create<TAction>(TAction action)
+            where TAction : IAction
+            => new DispatchContext<TAction>(action);
+    }
+}
