+++ src/StateR/IDispatchContextFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace StateR
+{
+    public interface IDispatchContextFactory
+    {
+        IDispatchContext<TAction> Create<TAction>(TAction action) where TAction : IAction;
+    }
+}
