+++ src/StateR/IDispatchManager.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace StateR
+{
+    public interface IDispatchManager
+    {
+        Task DispatchAsync<TAction>(IDispatchContext<TAction> dispatchContext, CancellationToken cancellationToken)
+            where TAction : IAction;
+    }
+}
