+++ src/StateR/Interceptors/IActionInterceptor.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace StateR.Interceptors
+{
+    public interface IActionInterceptor<TAction>
+        where TAction : IAction
+    {
+        Task InterceptAsync(IDispatchContext<TAction> context, CancellationToken cancellationToken);
+    }
+}
