+++ src/StateR/Interceptors/IInterceptorsManager.cs
@@ -1,0 +1,4 @@
+﻿namespace StateR.Interceptors
+{
+    public interface IInterceptorsManager : IDispatchManager { }
+}
