+++ src/StateR/Old/AsyncReducers/AsyncErrorOccured.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using StateR.Reducers;
+using System;
 using System.Collections;
 
 namespace StateR
