+++ src/StateR/Old/IStatorBuilder.cs
