+++ src/StateR/Old/IStore.cs
