+++ src/StateR/Old/StateHistoryEntry.cs
