+++ src/StateR/Old/Store.cs
