+++ src/StateR/Reducers/IActionHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace StateR
+namespace StateR.Reducers
 {
     public interface IActionHandler<TAction>
         where TAction : IAction
