+++ src/StateR/Reducers/IReducersManager.cs
@@ -1,0 +1,4 @@
+﻿namespace StateR.Reducers
+{
+    public interface IReducersManager : IDispatchManager { }
+}
